+++ MODULE_GLOB.mjs
@@ -3,6 +3,6 @@
 /**
  * File glob pattern to recursively match TypeScript (`.mts`, `.cts`, `.ts`, and
  * `.tsx`; excluding TypeScript definition files) and JavaScript (`.mjs`,
- * `.cjs`, and `.js`) modules.
+ * `.cjs`, `.js`, and `.jsx`) modules.
  */
-export default "**/{!(*.d).mts,!(*.d).cts,!(*.d).ts,*.{mjs,cjs,js,tsx}}";
+export default "**/{!(*.d).mts,!(*.d).cts,!(*.d).ts,*.{mjs,cjs,js,jsx,tsx}}";