+++ changelog.md
@@ -10,6 +10,10 @@
 
 - Enable JSX syntax when parsing `.jsx` files with Babel, even when the project has no Babel config for JSX, via [#8](https://github.com/jaydenseric/find-unused-exports/pull/8).
 
+### Patch
+
+- Improved the internal module `MODULE_GLOB` tests.
+
 ## 6.0.0
 
 ### Major